refact: move Loadable-related structures to the root package
