+++ loadable/src/main/java/com/jeanbarrossilva/loadable/LoadableScope.extensions.kt
@@ -8,7 +8,7 @@ import java.io.Serializable
  * @param send Callback run whenever a [Loadable] is sent.
  **/
 @Suppress("FunctionName")
-fun <T : Serializable?> LoadableScope(send: (Loadable<T>) -> Unit): LoadableScope<T> {
+internal fun <T : Serializable?> LoadableScope(send: (Loadable<T>) -> Unit): LoadableScope<T> {
     return object : LoadableScope<T>() {
         override suspend fun send(loadable: Loadable<T>) {
             send(loadable)
